requirements.txt
aiohttp==3.6.2
aniso8601==7.0.0
asgiref==3.2.2
asn1crypto==1.0.1
async-timeout==3.0.1
attrs==19.3.0
autobahn==19.10.1
Automat==0.8.0
bcrypt==3.1.7
cffi==1.12.3
channels==2.3.1
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
distro==1.4.0
Django==2.2.6
django-channels-graphql-ws==0.4.2
django-cors-middleware==1.4.0
django-graphql-jwt==0.2.2
ecdsa==0.13.3
future==0.17.1
graphene==2.1.8
graphene-django==2.6.0
graphene-file-upload==1.2.2
graphql-core==2.2.1
graphql-relay==2.0.0
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.2
multidict==4.5.2
numpy==1.17.3
pandas==0.25.2
Pebble==4.4.0
promise==2.2.1
psycopg2==2.8.3
pyasn1==0.4.7
pycparser==2.19
pycryptodome==3.9.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pyjwt-rsa==0.3.0
python-dateutil==2.8.0
pytz==2019.2
redis==3.3.11
rsa==4.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
tabula-py==1.4.1
Twisted==19.7.0
txaio==18.8.1
typing==3.7.4.1
yarl==1.3.0
zope.interface==4.6.0
